Overhaul and refactor of SoundContainers
Exposes more functionality for Sub Soundsets to control various
playback settings, which used to be a part of SoundContainer.

Refactor SoundContainer to not hold settings pertaining to how a
sound should be played to a SoundSettings class within SoundSets
instead.

Still need to check if this is exposed to lua or not though.